requirements.txt
Werkzeug==0.16.0
requests==2.18.4
Flask==1.1.1
flask_restplus==0.10.1
uwsgi==2.0.15
requests==2.18.4
pyotp==2.3.0
Pillow==6.2.1
numpy==1.17